requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pydantic>=2.0.0
